Fix card click status not updating
When the game restarts due to clicking the same card more than once,
the rest of the cards do not get their statuses changed back to "not
clicked". This commit fixes this by passing a new gameStatus state to
each card component. If it detects that the gameStatus is "Restarting",
it accordingly changes its state to "not clicked".